% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admissibility.R
\name{admissibility}
\alias{admissibility}
\title{Admissibility function (kappa) for one or multiple groups of constraints and block constraints}
\usage{
admissibility(state, constraint_list, log = TRUE)
}
\arguments{
\item{state}{a binary membership vector describing a feature set}

\item{constraint_list}{a list of constraint groups, each containing a matrix `A` and a vector `b` representing the inequality system `Ax<=b`, a vector `rho`, and a matrix `block_matrix`}

\item{log}{whether the admissibility should be returned on log scale}
}
\value{
an admissibility value
}
\description{
Evaluate the value of the admissibility function `kappa`.
}
