% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_admissibility.R
\name{block_admissibility}
\alias{block_admissibility}
\title{block admissibility function (kappa)}
\usage{
block_admissibility(state, constraints, log = TRUE)
}
\arguments{
\item{state}{a binary membership vector describing a feature set}

\item{constraints}{a list constaining a matrix A and a vector b representing the inequality system Ax<=b, a vector rho, and a block matrix}

\item{log}{whether the admissibility should be returned on log scale}
}
\value{
an admissibility value
}
\description{
evaluates the value of the admissibility function kappa for block constraints
}
