% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildConstraints.R
\name{build.UBayconstraint}
\alias{build.UBayconstraint}
\title{Build a customized constraint for UBayFS}
\usage{
build.UBayconstraint(A, b, rho, block_matrix = NULL)
}
\arguments{
\item{A}{matrix containing the left side of the linear inequality system}

\item{b}{vector containing the right side of the linear inequality system}

\item{rho}{vector containing the relaxation parameters for each constraint}

\item{block_matrix}{a matrix indicating the membership of features in feature blocks}
}
\value{
a `UBayconstraint` object
}
\description{
Builds a constraint using a left side `A`, a right side `b`, a relaxation parameter `rho`, and a block matrix `block_matrix`.
}
