% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildUBaymodel.R
\name{build.UBaymodel}
\alias{build.UBaymodel}
\title{Build an ensemble for UBayFS}
\usage{
build.UBaymodel(
  data,
  target,
  M = 100,
  tt_split = 0.75,
  nr_features = 10,
  method = "mRMR",
  constraints = NULL,
  block_constraints = NULL,
  weights = 1,
  constraint_dropout_rate = 0.1,
  popGreedy = 20,
  popsize = 50,
  maxiter = 100,
  shiny = FALSE
)
}
\arguments{
\item{data}{a matrix of input data}

\item{target}{a vector (factor) of labels for data}

\item{M}{the number of elementary models in the ensemble}

\item{tt_split}{the ratio of samples drawn for building an elementary model (train-test-split)}

\item{nr_features}{number of features to select in each elementary model}

\item{method}{a vector denoting the method(s) used as elementary models; options: "mRMR", "Laplacian score"}

\item{constraints}{a list containing a relaxed system Ax<=b of user constraints, given as matrix A, vector b and vector or scalar rho (relaxation parameters); see buildConstraints function}

\item{block_constraints}{a list containing a relaxed system Ax<=b of user constraints on feature blocks, given as matrix A, vector b and vector or scalar rho (relaxation parameters); see buildConstraints function}

\item{weights}{the vector of user-defined prior weights for each feature}

\item{constraint_dropout_rate}{rate of dropping constraints in Greedy algorithm}

\item{popGreedy}{size of the initial population obtained from Greedy algorithm. Must be smaller or equal to popsize.}

\item{popsize}{size of the initial population of the genetic algorithm for model optimization}

\item{maxiter}{maximum number of iterations of the genetic algorithm for model optimization}

\item{shiny}{TRUE indicates that the function is called from Shiny dashboard}
}
\value{
a UBaymodel object containing the following list elements: data, target, user.params (parameters representing user knowledge), ensemble.params (parameters representing the likelihood) and optim.params (parameters for genetic algorithm)
}
\description{
builds a data structure for UBayFS and trains an ensemble of elementary feature selectors
}
\details{
The function aggregates input parameters for UBayFS - including data, parameters defining ensemble and user knowledge and parameters specifying the optimization procedure - and trains the ensemble model
}
\examples{
# build a UBayFS model using Wisconsin breast cancer dataset
data(wbc) # dataset
c <- buildConstraints(constraint_types = "max_size",
                      constraint_vars = list(10),
                      num_elements = ncol(wbc$data),
                      rho = 1) # prior constraints
w <- rep(1, ncol(wbc$data)) # weights
model <- build.UBaymodel(
                     data = wbc$data,
                     target = wbc$labels,
                     constraints = c,
                     weights = w
)

# include block-constraints
c_block <- buildConstraints(constraint_types = "max_size",
                            constraint_vars = list(2),
                            num_elements = length(wbc$blocks),
                            rho = 10,
                            block_list = wbc$blocks)
model <- build.UBaymodel(
                     data = wbc$data,
                     target = wbc$labels,
                     constraints = c,
                     block_constraints = c_block,
                     weights = w
)
}
