% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildUBaymodel.R
\name{build_train_set}
\alias{build_train_set}
\title{Perform stratified data partition.}
\usage{
build_train_set(y, tt_split)
}
\arguments{
\item{y}{a column, often the target, by which the data shall be partitioned.}

\item{tt_split}{the percentage of data used for training in each ensemble model.}
}
\value{
data indices for training ensembles
}
\description{
Sample indices for training from the data.
}
