% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateFS.R
\name{evaluateFS}
\alias{evaluateFS}
\title{evaluate feature set}
\usage{
evaluateFS(state, model, method = "spearman", log = TRUE)
}
\arguments{
\item{state}{a binary membership vector describing a feature set}

\item{model}{a UBaymodel object created using build.UBaymodel}

\item{log}{whether the admissibility should be returned on log scale}
}
\value{
a posterior probability value
}
\description{
evaluates a feature set under the UBayFS model framework
}
