% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPosterior.R
\name{getPosterior}
\alias{getPosterior}
\title{Model posterior}
\usage{
getPosterior(state, model, log = TRUE)
}
\arguments{
\item{state}{a binary membership vector describing a feature set}

\item{model}{a UBaymodel object created using build.UBaymodel}

\item{log}{whether the admissibility should be returned on log scale}
}
\value{
a posterior probability value
}
\description{
evaluates the value of the posterior probability density (target function)
}
