% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResults.R
\name{getResults}
\alias{getResults}
\title{feature selection results}
\usage{
getResults(model)
}
\arguments{
\item{model}{a UBaymodel object created using build.UBaymodel after training}
}
\description{
displays and summarizes the results of UBayFS after feature selection.
}
