% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admissibility.R
\name{group_admissibility}
\alias{group_admissibility}
\title{Admissibility for constraint group}
\usage{
group_admissibility(state, constraints, log = TRUE)
}
\description{
computes admissibility for a group of constraints (with a common block)
}
