% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildConstraints.R
\name{is.UBayconstraint}
\alias{is.UBayconstraint}
\title{Checks whether a list object implements proper UBayFS user constraints}
\usage{
is.UBayconstraint(x)
}
\arguments{
\item{x}{a `UBayconstraint` object}
}
\value{
boolean value
}
\description{
Checks whether a list object implements proper UBayFS user constraints
}
