% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isUBaymodel.R
\name{is.UBaymodel}
\alias{is.UBaymodel}
\title{checks whether an object is a UBaymodel}
\usage{
is.UBaymodel(x)
}
\arguments{
\item{x}{an object to be checked for class consistency}
}
\description{
perform consistency checks of a UBaymodel
}
