% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadWisconsin.R
\name{load_wisconsin}
\alias{load_wisconsin}
\title{loads the Wisconsin breast cancer dataset remotely}
\usage{
load_wisconsin()
}
\value{
a list with elements data and labels
}
\description{
loads the Wisconsin breast cancer dataset as demonstration dataset from UCI database
}
