% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadWisconsin.R
\name{loadWisconsin}
\alias{loadWisconsin}
\title{loads the Wisconsin breast cancer dataset remotely}
\usage{
loadWisconsin()
}
\value{
a list with elements data and labels
}
\description{
loads the Wisconsin breast cancer dataset as demonstration dataset from UCI database
}
