% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPosterior.R
\name{posterior}
\alias{posterior}
\title{posterior function}
\usage{
posterior(state, post_param, log = TRUE)
}
\arguments{
\item{state}{a binary membership vector describing a feature set}

\item{post_param}{a vector of posterior weights for the Dirichlet distribution}

\item{log}{whether the admissibility should be returned on log scale}

\item{model}{a UBaymodel object created using build.UBaymodel}
}
\value{
a posterior probability value
}
\description{
evaluates the value of the posterior probability density (target function)
}
