% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admissibility.R
\name{posteriorExpectation}
\alias{posteriorExpectation}
\title{Posterior expectation of features}
\usage{
posteriorExpectation(model)
}
\arguments{
\item{model}{a `UBaymodel` object}
}
\description{
compute the posterior score for each feature.
}
