% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printModel.R
\name{print.UBaymodel}
\alias{print.UBaymodel}
\title{prints a UBayFS model}
\usage{
\method{print}{UBaymodel}(model)
}
\arguments{
\item{model}{a UBaymodel object created using build.model}
}
\description{
print details of an UBayFS model
}
