% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runInteractive.R
\name{runInteractive}
\alias{runInteractive}
\title{Run an interactive Shiny app for demonstration}
\usage{
runInteractive()
}
\value{
calls Shiny app, no return value
}
\description{
Starts an interactive R Shiny application in the browser.
}
