% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleInitial.R
\name{sampleInitial}
\alias{sampleInitial}
\title{initial feature set sampling using probabilistic Greedy algorithm}
\usage{
sampleInitial(
  post_scores,
  constraints,
  block_constraints,
  size,
  constraint_dropout_rate = NULL
)
}
\arguments{
\item{post_scores}{a vector of posterior scores (prior scores + likelihood) for each feature}

\item{constraints}{a list containing feature-wise constraints}

\item{block_constraints}{a list containing block-wise constraints}

\item{size}{initial number of samples to be created. The output sample size can be lower, since duplicates are removed.}

\item{constraint_dropout_rate}{rate of dropping constraints in Greedy algorithm}
}
\value{
a matrix containing initial feature sets as rows.
}
\description{
sample initial solutions using a probabilistic version of Greedy algorithm
}
