% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setConstraintParams.R
\name{set_constraint_params}
\alias{set_constraint_params}
\title{set constraints in UBaymodel object}
\usage{
set_constraint_params(model, A, b, rho)
}
\arguments{
\item{model}{a UBaymodel object created using build.model}

\item{A}{the matrix defining the constraint system Ax<=b}

\item{b}{the vector defining the constraint system Ax<=b}

\item{rho}{the vector of relaxation parameters for the constraint system Ax<=b}
}
\value{
a UBaymodel object with updated constraint parameters
}
\description{
sets the constraints in a UBaymodel object
}
\seealso{
build.model
}
