% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setOptimParams.R
\name{set_optim_params}
\alias{set_optim_params}
\title{set optimization parameters in UBaymodel object}
\usage{
set_optim_params(model, popsize, maxiter)
}
\arguments{
\item{model}{a UBaymodel object created using build.model}

\item{popsize}{size of the initial population of the genetic algorithm for model optimization}

\item{maxiter}{maximum number of iterations of the genetic algorithm for model optimization}
}
\value{
a UBaymodel object with updated optimization parameters
}
\description{
sets the optimization parameters in a UBaymodel object
}
\seealso{
build.model
}
