% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setWeightParams.R
\name{set_weight_params}
\alias{set_weight_params}
\title{set weights in UBaymodel object}
\usage{
set_weight_params(model, weights)
}
\arguments{
\item{model}{a UBaymodel object created using build.model}

\item{weights}{the vector of user-defined prior weights for each feature}
}
\value{
a UBaymodel object with updated prior weights
}
\description{
sets the prior weights in a UBaymodel object
}
\seealso{
build.model
}
