% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setWeights.R
\name{setWeights}
\alias{setWeights}
\title{set weights in UBaymodel object}
\usage{
setWeights(model, weights)
}
\arguments{
\item{model}{a UBaymodel object created using build.UBaymodel}

\item{weights}{the vector of user-defined prior weights for each feature}
}
\value{
a UBaymodel object with updated prior weights
}
\description{
sets the prior weights in a UBaymodel object
}
\seealso{
build.UBaymodel
}
