% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainUBaymodel.R
\name{train}
\alias{train}
\title{UBayFS feature selection}
\usage{
train(x)
}
\arguments{
\item{x}{a UBaymodel created by build.UBaymodel}
}
\value{
A UBaymodel with an additional list element output containing the maximum a-posteriori estimate (map)
}
\description{
Genetic algorithm to train UBayFS feature selection model.
}
