% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainModel.R
\name{train_model}
\alias{train_model}
\title{perform UBayFS feature selection}
\usage{
train_model(model)
}
\arguments{
\item{model}{a UBaymodel created by build.model}
}
\value{
a UBaymodel with an additional list element output containing the maximum a-posteriori estimate (map)
}
\description{
runs a genetic algorithm to perform UBayFS feature selection
}
