% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admissibility.R
\name{admissibility}
\alias{admissibility}
\title{admissibility function (kappa)}
\usage{
admissibility(state, A, b, rho, weights_sum = NULL, log = TRUE)
}
\arguments{
\item{state}{a binary membership vector describing a feature set}

\item{A}{the matrix defining the constraint system Ax<=b}

\item{b}{the vector defining the constraint system Ax<=b}

\item{rho}{the vector of relaxation parameters for the constraint system Ax<=b}

\item{weights_sum}{a scalar denoting the sum of all feature weights to scale rho; if NULL, no weighting is performed}

\item{log}{whether the admissibility should be returned on log scale}
}
\value{
an admissibility value
}
\description{
evaluates the value of the admissibility function kappa
}
