% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admissibility.R
\name{admissibility}
\alias{admissibility}
\title{admissibility function (kappa)}
\usage{
admissibility(state, constraints, weights_sum = NULL, log = TRUE)
}
\arguments{
\item{state}{a binary membership vector describing a feature set}

\item{constraints}{a list constaining a matrix A and a vector b representing the inequality system Ax<=b and a vector rho}

\item{weights_sum}{a scalar denoting the sum of all feature weights to scale rho; if NULL, no weighting is performed}

\item{log}{whether the admissibility should be returned on log scale}
}
\value{
an admissibility value
}
\description{
evaluates the value of the admissibility function kappa
}
