% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admissibility.R
\name{group_admissibility}
\alias{group_admissibility}
\alias{admissibility}
\title{Admissibility for constraint group}
\usage{
group_admissibility(state, constraints, log = TRUE)

admissibility(state, constraint_list, log = TRUE)
}
\arguments{
\item{state}{a binary membership vector describing a feature set}

\item{constraints}{group of constraints with common block matrix}

\item{log}{whether the admissibility should be returned on log scale}

\item{constraint_list}{a list of constraint groups, each containing a matrix `A` and a vector `b` representing the inequality system `Ax<=b`, a vector `rho`, and a matrix `block_matrix`}
}
\value{
an admissibility value
}
\description{
Evaluate the value of the admissibility function `kappa`.
}
\section{Functions}{
\itemize{
\item \code{group_admissibility()}: computes admissibility for a group of constraints (with a common block).

}}
