% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_admissibility.R
\name{block_admissibility}
\alias{block_admissibility}
\title{block admissibility function (kappa)}
\usage{
block_admissibility(
  state,
  A,
  b,
  rho,
  block_matrix,
  weights_sum = NULL,
  log = TRUE
)
}
\arguments{
\item{state}{a binary membership vector describing a feature set}

\item{A}{the matrix defining the constraint system Ax<=b between blocks}

\item{b}{the vector defining the constraint system Ax<=b between blocks}

\item{rho}{the vector of relaxation parameters for the constraint system Ax<=b between blocks}

\item{block_matrix}{the matrix containing affiliations of features to each block}

\item{weights_sum}{a scalar denoting the sum of all block weights to scale rho; if NULL, no weighting is performed}

\item{log}{whether the admissibility should be returned on log scale}
}
\value{
an admissibility value
}
\description{
evaluates the value of the admissibility function kappa for block constraints
}
