% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateFS.R
\name{evaluateFS}
\alias{evaluateFS}
\alias{evaluateMultiple}
\title{Evaluate  a feature set}
\usage{
evaluateFS(state, model, method = "spearman", log = FALSE)

evaluateMultiple(state, model, method = "spearman", log = TRUE)
}
\arguments{
\item{state}{a binary membership vector describing a feature set}

\item{model}{a `UBaymodel` object created using \link{build.UBaymodel}}

\item{method}{type of correlation (`pearson`,`kendall`, or `spearman`)}

\item{log}{whether the admissibility should be returned on log scale}
}
\value{
a posterior probability value
}
\description{
Evaluates a feature set under the UBayFS model framework.
}
\section{Functions}{
\itemize{
\item \code{evaluateMultiple()}: Evaluate multiple feature sets

}}
