% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildUBaymodel.R
\name{is.UBaymodel}
\alias{is.UBaymodel}
\title{Check whether an object is a UBaymodel}
\usage{
is.UBaymodel(x)
}
\arguments{
\item{x}{an object to be checked for class consistency}
}
\value{
returns a single scalar (TRUE or FALSE) indicating whether the object fulfills the consistency requirements of the UBayFS model
}
\description{
Perform consistency checks of a UBaymodel.
}
