% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLikelihood.R
\name{likelihood}
\alias{likelihood}
\title{likelihood}
\usage{
likelihood(state, param, dmax, log = TRUE)
}
\arguments{
\item{state}{a binary membership vector describing a feature set}

\item{param}{a vector of posterior weights for the Multinomial function}

\item{dmax}{an positive integer denoting the maximum degree of feature interactions considered in the model}

\item{log}{whether the admissibility should be returned on log scale}
}
\value{
a posterior probability value
}
\description{
evaluates the value of the model likelihood
}
