% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUBaymodel.R
\name{plot.UBaymodel}
\alias{plot.UBaymodel}
\title{plot a UBayFS model}
\usage{
\method{plot}{UBaymodel}(x, ...)
}
\arguments{
\item{x}{a UBaymodel object created using build.UBaymodel}

\item{...}{unused}
}
\description{
a barplot of an UBayFS model containing prior weights, ensemble counts and the selected features
}
