% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPosterior.R
\name{posterior}
\alias{posterior}
\title{posterior function}
\usage{
posterior(
  state,
  constraints,
  block_constraints,
  weights,
  counts,
  dmax,
  log = TRUE
)
}
\arguments{
\item{state}{a binary membership vector describing a feature set}

\item{constraints}{a list constaining a matrix A and a vector b representing the inequality system Ax<=b and a vector rho}

\item{block_constraints}{a list containing a relaxed system Ax<=b of user constraints on feature blocks, given as matrix A, vector b and vector or scalar rho (relaxation parameters); see buildConstraints function}

\item{weights}{a vector of prior weights for the Dirichlet distribution}

\item{counts}{a vector of likelihood counts for the Multinomial distribution}

\item{dmax}{an positive integer denoting the maximum degree of feature interactions considered in the model}

\item{log}{whether the admissibility should be returned on log scale}
}
\value{
a posterior probability value
}
\description{
evaluates the value of the posterior probability density (target function)
}
