% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printUBaymodel.R
\name{print.UBaymodel}
\alias{print.UBaymodel}
\title{prints a UBayFS model}
\usage{
\method{print}{UBaymodel}(x, ...)
}
\arguments{
\item{x}{a UBaymodel object created using build.UBaymodel}

\item{...}{unused}
}
\description{
print details of an UBayFS model
}
