% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainUBaymodel.R
\name{sampleInitial}
\alias{sampleInitial}
\title{Initial feature set sampling using probabilistic Greedy algorithm}
\usage{
sampleInitial(post_scores, constraints, size)
}
\arguments{
\item{post_scores}{a vector of posterior scores (prior scores + likelihood) for each feature}

\item{constraints}{a list containing feature-wise constraints}

\item{size}{initial number of samples to be created. The output sample size can be lower, since duplicates are removed.}
}
\value{
a matrix containing initial feature sets as rows
}
\description{
Sample initial solutions using a probabilistic version of Greedy algorithm.
}
