% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildConstraints.R
\name{setBlockConstraints}
\alias{setBlockConstraints}
\title{Set block constraints in UBaymodel object.}
\usage{
setBlockConstraints(model, constraints, append = FALSE)
}
\arguments{
\item{model}{a UBaymodel object created using build.UBaymodel}

\item{constraints}{a list containing a relaxed system Ax<=b of user constraints on feature blocks, given as matrix A, vector b and vector or scalar rho (relaxation parameters); see buildConstraints function}

\item{append}{whether the constraint should be appended to previous constraints or overwrite...}
}
\value{
a UBaymodel object with updated constraint parameters
}
\description{
sets the block constraints in a UBaymodel object
}
\seealso{
build.UBaymodel
}
