% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setConstraints.R
\name{setConstraints}
\alias{setConstraints}
\title{set constraints in UBaymodel object}
\usage{
setConstraints(model, constraints, append = FALSE)
}
\arguments{
\item{model}{a UBaymodel object created using build.UBaymodel}

\item{constraints}{a list containing a relaxed system Ax<=b of user constraints, given as matrix A, vector b and vector or scalar rho (relaxation parameters); see buildConstraints function}

\item{append}{if TRUE, constraints are appended to the existing constraint system}
}
\value{
a UBaymodel object with updated constraint parameters
}
\description{
sets the constraints in a UBaymodel object
}
\seealso{
build.UBaymodel
}
