% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildConstraints.R
\name{setConstraints}
\alias{setConstraints}
\alias{setBlockConstraints}
\title{Set constraints in UBaymodel object}
\usage{
setConstraints(model, constraints, append = FALSE)

setBlockConstraints(model, constraints, append = FALSE)
}
\arguments{
\item{model}{a `UBaymodel` object created using \link{build.UBaymodel}}

\item{constraints}{a `UBayconstraint` object created using \link{build.UBayconstraint}}

\item{append}{if `TRUE`, constraints are appended to the existing constraint system}
}
\value{
a `UBaymodel` object with updated constraint parameters
}
\description{
Set the constraints in a `UBaymodel` object.

Set the block constraints in a `UBaymodel` object.
}
\section{Functions}{
\itemize{
\item \code{setBlockConstraints()}: sets the block constraints in a `UBaymodel` object

}}
\seealso{
build.UBaymodel
}
