% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryModel.R
\name{summary.UBaymodel}
\alias{summary.UBaymodel}
\title{summarizes a UBayFS model}
\usage{
\method{summary}{UBaymodel}(object, ...)
}
\arguments{
\item{object}{a UBaymodel object created using build.model}

\item{...}{unused}
}
\description{
a summary of an UBayFS model
}
